+++ tap_eloqua/discover.py
@@ -1,9 +1,11 @@
 from requests.exceptions import HTTPError
+import singer
 from singer.catalog import Catalog, CatalogEntry, Schema
 
 from tap_eloqua.schema import get_schemas, get_pk
-from tap_eloqua.utils import check_stream_access
 from tap_eloqua.constants import STATIC_ENDPOINTS
+
+LOGGER = singer.get_logger()
 
 # Derive probe paths from the shared STATIC_ENDPOINTS constant.
 # Each path is prefixed with the REST API base to form a full probe URL.
@@ -11,6 +13,34 @@ STATIC_STREAM_PROBE_PATHS = {
     ep['stream_id']: f"/api/REST/2.0/{ep['path']}"
     for ep in STATIC_ENDPOINTS
 }
+
+
+def check_stream_access(stream_name, probe_fn, auth_error_types, fallback_accessible=False):
+    """
+    Probe a stream endpoint and return True if accessible, False on auth error.
+
+    :param stream_name: Used in log messages.
+    :param probe_fn: Zero-argument callable that performs the API probe.
+    :param auth_error_types: Exception type(s) indicating 401/403 — returns False.
+    :param fallback_accessible: If True, non-auth errors (e.g. 400 from minimal
+                                probe params) are treated as auth-OK and return True.
+                                If False (default), they are re-raised.
+    """
+    try:
+        probe_fn()
+        LOGGER.info("Stream '%s' is accessible.", stream_name)
+        return True
+    except auth_error_types:
+        LOGGER.warning(
+            "Stream '%s' is not accessible with the provided credentials.",
+            stream_name,
+        )
+        return False
+    except Exception:  # pylint: disable=broad-except
+        if fallback_accessible:
+            LOGGER.info("Stream '%s' endpoint reachable (auth OK).", stream_name)
+            return True
+        raise
 
 
 def _is_auth_http_error(exc):
@@ -66,5 +96,11 @@ def discover(client):
             schema=schema,
             metadata=mdata
         ))
+
+    if not catalog.streams:
+        raise Exception(
+            "No stream endpoints are accessible with the provided credentials. "
+            "Verify that the API credentials have the required permissions."
+        )
 
     return catalog
